Run as a server with nanomsg communication
Running as a flask app was limiting in that it was impossible(?) to
maintain python objects for the entire runtime of the application. In
turn, this made it impossible to add for example an external GPIO button
to the app. The new nanomessage server architecture makes it easy to add
such functionality and it also made it easy to do some setup when
starting, for example bringing the session LED to the correct state.